requirements.txt
# Análisis de datos
pandas==2.0.3
numpy==1.24.3

# Machine Learning
scikit-learn==1.3.0

# MLflow y tracking
mlflow==2.5.0

# Aplicación web
streamlit==1.25.0

# Visualización
plotly==5.15.0

# Configuración
python-dotenv==1.0.0

# Desarrollo
jupyter==1.0.0

# Testing (opcional)
pytest==7.4.0
pytest-cov==4.1.0